Bosh link preparatory manifest changes
As per: https://github.com/cloudfoundry/cf-release/releases/tag/v253
This will allow the consul job to start providing a bosh link without
having that link be consumed by the various consul jobs in the
deployment. CF 254 will introduce a version of consul that requires
these changes to the manifest, so please ensure that you deploy CF 253
and Diego 1.8.1 first before moving on to CF 254.